+++ src/main/java/module-info.java
@@ -21,4 +21,6 @@ module david.buzas.whattoeat {
     opens david.buzas.whattoeat.repositories to javafx.fxml;
     exports david.buzas.whattoeat.controllers;
     opens david.buzas.whattoeat.controllers to javafx.fxml;
+    exports david.buzas.whattoeat.utils;
+    opens david.buzas.whattoeat.utils to javafx.fxml;
 }
